tui/pyproject.toml
[project]
name = "jamarr-tui"
version = "0.1.0"
description = "Terminal UI client for Jamarr"
readme = "README.md"
license = "AGPL-3.0-only"
requires-python = ">=3.14"
dependencies = [
    "textual>=0.85",
    "httpx>=0.27",
    "pillow>=10.0",
]

[project.scripts]
jamarr-tui = "jamarr_tui.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["jamarr_tui"]
